requirements.txt
flask
python-dotenv
openai
aiohttp
requests
openai-whisper
torch
torchvision
torchaudio
ffmpeg
ffmpeg-python